Matrix data table added

+++ src/main/kotlin/raytracer/math/Constants.kt
@@ -3,11 +3,12 @@ package raytracer.math
 import kotlin.math.abs
 
 const val PI            = 3.1415926536f
-const val EPSILON       = 0.00001
-const val EPSILON_TEST  = 0.0001
+const val EPSILON       = 0.00001f
+const val EPSILON_TEST  = 0.0001f
 
-inline infix fun Double.eq(rhs: Double): Boolean =
+inline infix fun Float.eq(rhs: Float): Boolean =
     abs(this - rhs) <= EPSILON
 
-inline infix fun Double.neq(rhs: Double): Boolean =
+inline infix fun Float.neq(rhs: Float): Boolean =
     abs(this - rhs) > EPSILON
+
